linera-bridge: expose BurnBlocked test helpers for downstream tests
